auth added

+++ src/actions/actions.ts
@@ -2,28 +2,48 @@
 
 import prisma from "@/lib/db";
 import { sleep } from "@/lib/utils";
-import { petFormSchema, petIdSchema } from "@/lib/validations";
+import { authSchema, petFormSchema, petIdSchema } from "@/lib/validations";
 import { signIn, signOut } from "@/lib/auth";
 import { revalidatePath } from "next/cache";
 import { redirect } from "next/navigation";
 import bcrypt from "bcryptjs";
+import { checkAuth, getPetById } from "@/lib/server-utils";
+import { get } from "http";
 
 // --- User actions ---
 
-export async function logIn(formData: FormData) {
+export async function logIn(formData: unknown) {
+  //check if formData is a FormData type
+  if (!(formData instanceof FormData)) {
+    return { message: "Invalid form data" };
+  }
+
   await signIn("credentials", formData);
   redirect("/app/dashboard");
 }
 
-export async function signUp(formData: FormData) {
-  const hashedPassword = await bcrypt.hash(
-    formData.get("password") as string,
-    10
-  );
+export async function signUp(formData: unknown) {
+  //check if formData is a FormData type
+  if (!(formData instanceof FormData)) {
+    return { message: "Invalid form data" };
+  }
+
+  //convert formData to a plain object
+  const formDataEntries = Object.fromEntries(formData.entries());
+
+  //validation
+  const validatedFormData = authSchema.safeParse(formDataEntries);
+  if (!validatedFormData.success) {
+    return { message: "Invalid form data" };
+  }
+
+  const { email, password } = validatedFormData.data;
+
+  const hashedPassword = await bcrypt.hash(password, 10);
 
   await prisma.user.create({
     data: {
-      email: formData.get("email") as string,
+      email,
       hashedPassword,
     },
   });
@@ -40,6 +60,8 @@ export async function logOut() {
 export async function addPet(pet: unknown) {
   await sleep(1000);
 
+  const session = await checkAuth();
+
   const validatedPet = petFormSchema.safeParse(pet);
   if (!validatedPet.success) {
     return { message: "Invalid pet data" };
@@ -47,7 +69,14 @@ export async function addPet(pet: unknown) {
 
   try {
     await prisma.pet.create({
-      data: validatedPet.data,
+      data: {
+        ...validatedPet.data,
+        user: {
+          connect: {
+            id: session.user.id,
+          },
+        },
+      },
     });
   } catch (error) {
     return { message: "Failed to add pet" };
@@ -58,12 +87,28 @@ export async function addPet(pet: unknown) {
 export async function editPet(petId: unknown, newPetData: unknown) {
   await sleep(1000);
 
+  //authentication check
+  const session = await checkAuth();
+
+  //validation
   const validatedPetId = petIdSchema.safeParse(petId);
   const validatedPet = petFormSchema.safeParse(newPetData);
 
   if (!validatedPetId.success || !validatedPet.success) {
     return { message: "Invalid pet data" };
   }
+
+  //authorization check (user can only edit their own pets)
+  const pet = await getPetById(validatedPetId.data);
+
+  if (!pet) {
+    return { message: "Pet not found" };
+  }
+  if (pet.userId !== session.user.id) {
+    return { message: "Unauthorized" };
+  }
+
+  //database mutation
   try {
     await prisma.pet.update({
       where: {
@@ -80,11 +125,26 @@ export async function editPet(petId: unknown, newPetData: unknown) {
 export async function deletePet(petId: unknown) {
   await sleep(1000);
 
-  const validatedPetId = petIdSchema.safeParse(petId);
+  //authentication check
+  const session = await checkAuth();
 
+  //validation
+  const validatedPetId = petIdSchema.safeParse(petId);
   if (!validatedPetId.success) {
     return { message: "Invalid pet data" };
   }
+
+  //authorization check (user can only delete their own pets)
+  const pet = await getPetById(validatedPetId.data);
+
+  if (!pet) {
+    return { message: "Pet not found" };
+  }
+  if (pet.userId !== session.user.id) {
+    return { message: "Unauthorized" };
+  }
+
+  //database mutation
   try {
     await prisma.pet.delete({
       where: {
